feat(usuarios): alinhar layout de header e formulários ao padrão Athos
