+++ apps/usuarios/templates/usuarios/componentes/filtros.html
@@ -1,4 +1,4 @@
-<form method="get" class="rounded-2xl border border-gray-200 bg-white p-4 shadow-sm">
+<form method="get" class="rounded-2xl border border-gray-200 bg-white/95 p-4 shadow-sm">
     <div class="flex flex-col gap-4 lg:flex-row lg:items-center">
         <div class="relative flex-1">
             <span class="pointer-events-none absolute inset-y-0 left-4 flex items-center text-gray-400">
@@ -20,11 +20,11 @@
         </div>
         <div class="flex items-center gap-2 lg:ml-auto">
             <a href="{% url 'usuarios:lista' %}"
-               class="inline-flex items-center gap-2 rounded-full border border-gray-200 px-4 py-2 text-sm font-medium text-gray-600 hover:bg-gray-100">
+               class="inline-flex items-center gap-2 rounded-xl border border-gray-200 px-4 py-2 text-sm font-medium text-gray-600 transition hover:border-gray-300 hover:bg-gray-100">
                 Limpar
             </a>
             <button type="submit"
-                    class="inline-flex items-center gap-2 rounded-full bg-blue-600 px-5 py-2 text-sm font-semibold text-white hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-blue-500 focus:ring-offset-2">
+                    class="inline-flex items-center gap-2 rounded-xl bg-gradient-to-r from-indigo-500 via-purple-500 to-pink-500 px-5 py-2 text-sm font-semibold text-white shadow-md transition hover:brightness-105 focus:outline-none focus:ring-2 focus:ring-purple-200 focus:ring-offset-1">
                 <i class="fas fa-filter"></i>
                 Aplicar
             </button>
