+++ apps/usuarios/templates/usuarios/componentes/form_usuario.html
@@ -1,4 +1,4 @@
-<form method="post" class="rounded-3xl border border-gray-100 bg-white p-8 shadow-xl space-y-6">
+<form method="post" class="space-y-5">
     {% csrf_token %}
     {% if form.non_field_errors %}
         <div class="rounded-md border border-red-200 bg-red-50 px-4 py-3 text-sm text-red-700">
@@ -8,7 +8,7 @@
         </div>
     {% endif %}
 
-    <div class="grid gap-5 md:grid-cols-2">
+    <div class="grid gap-4 md:grid-cols-2">
         <div class="md:col-span-2">
             <label for="{{ form.nome_completo.id_for_label }}" class="text-sm font-medium text-gray-700">
                 {{ form.nome_completo.label }} <span class="text-red-500">*</span>
@@ -40,11 +40,14 @@
         </div>
 
         <div>
-            <span class="text-sm font-medium text-gray-700">
+            <label for="{{ form.contrato.id_for_label }}" class="text-sm font-medium text-gray-700">
                 {{ form.contrato.label }} <span class="text-red-500">*</span>
-            </span>
-            <div class="mt-2 space-y-2 text-sm text-gray-700">
+            </label>
+            <div class="relative mt-1">
                 {{ form.contrato }}
+                <span class="pointer-events-none absolute inset-y-0 right-3 flex items-center text-gray-400">
+                    <i class="fas fa-chevron-down text-xs"></i>
+                </span>
             </div>
             {% for error in form.contrato.errors %}
                 <p class="mt-1 text-xs text-red-600">{{ error }}</p>
@@ -61,30 +64,38 @@
             {% endfor %}
         </div>
 
-        <div class="md:col-span-2">
-            <span class="text-sm font-medium text-gray-700">
+        <div class="md:col-span-1">
+            <label for="{{ form.perfil_acesso.id_for_label }}" class="text-sm font-medium text-gray-700">
                 {{ form.perfil_acesso.label }} <span class="text-red-500">*</span>
-            </span>
-            <div class="mt-2 space-y-2 text-sm text-gray-700">
+            </label>
+            <div class="relative mt-1">
                 {{ form.perfil_acesso }}
+                <span class="pointer-events-none absolute inset-y-0 right-3 flex items-center text-gray-400">
+                    <i class="fas fa-chevron-down text-xs"></i>
+                </span>
             </div>
             {% for error in form.perfil_acesso.errors %}
                 <p class="mt-1 text-xs text-red-600">{{ error }}</p>
             {% endfor %}
         </div>
 
-        <div class="md:col-span-2">
-            <label class="inline-flex items-center gap-2 text-sm font-medium text-gray-700">
-                {{ form.ativo }}
-                <span>Usuário ativo</span>
+        <div class="md:col-span-1">
+            <label for="{{ form.ativo.id_for_label }}" class="text-sm font-medium text-gray-700">
+                {{ form.ativo.label }}
             </label>
+            <div class="relative mt-1">
+                {{ form.ativo }}
+                <span class="pointer-events-none absolute inset-y-0 right-3 flex items-center text-gray-400">
+                    <i class="fas fa-chevron-down text-xs"></i>
+                </span>
+            </div>
             {% for error in form.ativo.errors %}
                 <p class="mt-1 text-xs text-red-600">{{ error }}</p>
             {% endfor %}
         </div>
 
         {% if form.fields.senha %}
-            <div class="md:col-span-2">
+            <div>
                 <label for="{{ form.senha.id_for_label }}" class="text-sm font-medium text-gray-700">
                     {{ form.senha.label }} <span class="text-red-500">*</span>
                 </label>
@@ -96,7 +107,7 @@
                     <p class="mt-1 text-xs text-red-600">{{ error }}</p>
                 {% endfor %}
             </div>
-            <div class="md:col-span-2">
+            <div>
                 <label for="{{ form.confirmar_senha.id_for_label }}" class="text-sm font-medium text-gray-700">
                     {{ form.confirmar_senha.label }} <span class="text-red-500">*</span>
                 </label>
@@ -111,7 +122,7 @@
         {% endif %}
 
         {% if form.fields.nova_senha %}
-            <div class="md:col-span-2">
+            <div>
                 <label for="{{ form.nova_senha.id_for_label }}" class="text-sm font-medium text-gray-700">
                     {{ form.nova_senha.label }}
                 </label>
@@ -123,7 +134,7 @@
                     <p class="mt-1 text-xs text-red-600">{{ error }}</p>
                 {% endfor %}
             </div>
-            <div class="md:col-span-2">
+            <div>
                 <label for="{{ form.confirmar_nova_senha.id_for_label }}" class="text-sm font-medium text-gray-700">
                     {{ form.confirmar_nova_senha.label }}
                 </label>
@@ -151,11 +162,11 @@
 
     <div class="flex justify-end gap-3">
         <a href="{% url 'usuarios:lista' %}"
-           class="inline-flex items-center gap-2 rounded-full border border-gray-300 px-5 py-2 text-sm font-medium text-gray-600 hover:bg-gray-100">
+           class="inline-flex items-center gap-2 rounded-xl border border-gray-300 px-5 py-2 text-sm font-medium text-gray-600 transition hover:border-gray-400 hover:bg-gray-100">
             Cancelar
         </a>
         <button type="submit"
-                class="inline-flex items-center gap-2 rounded-full bg-blue-600 px-6 py-2 text-sm font-semibold text-white hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-blue-500 focus:ring-offset-2">
+                class="inline-flex items-center gap-2 rounded-xl bg-gradient-to-r from-indigo-500 via-purple-500 to-pink-500 px-6 py-2 text-sm font-semibold text-white shadow-md transition hover:brightness-105 focus:outline-none focus:ring-2 focus:ring-purple-200 focus:ring-offset-1">
             <i class="fas fa-save"></i>
             {{ submit_label|default:"Salvar" }}
         </button>
