+++ apps/usuarios/templates/usuarios/componentes/lista.html
@@ -1,16 +1,16 @@
-<div class="bg-white rounded-lg shadow overflow-hidden">
+<div class="overflow-hidden rounded-2xl border border-gray-200 bg-white shadow">
     <div class="overflow-x-auto">
         <table class="min-w-full divide-y divide-gray-200 text-sm">
             <thead class="bg-gray-50">
                 <tr class="text-left text-xs font-semibold uppercase tracking-wider text-gray-500">
-                    <th class="px-4 py-3">Nome</th>
-                    <th class="px-4 py-3">Username</th>
-                    <th class="px-4 py-3">E-mail</th>
-                    <th class="px-4 py-3">Cargo</th>
-                    <th class="px-4 py-3">Perfil de Acesso</th>
-                    <th class="px-4 py-3">Status</th>
-                    <th class="px-4 py-3">Data de Registro</th>
-                    <th class="px-4 py-3 text-right">Ações</th>
+                    <th class="px-4 py-3 text-slate-600">Nome</th>
+                    <th class="px-4 py-3 text-slate-600">Username</th>
+                    <th class="px-4 py-3 text-slate-600">E-mail</th>
+                    <th class="px-4 py-3 text-slate-600">Cargo</th>
+                    <th class="px-4 py-3 text-slate-600">Perfil de Acesso</th>
+                    <th class="px-4 py-3 text-slate-600">Status</th>
+                    <th class="px-4 py-3 text-slate-600">Data de Registro</th>
+                    <th class="px-4 py-3 text-right text-slate-600">Ações</th>
                 </tr>
             </thead>
             <tbody class="bg-white divide-y divide-gray-100 text-gray-700">
@@ -26,7 +26,7 @@
                         <td class="px-4 py-3">{{ usuario.cargo }}</td>
                         <td class="px-4 py-3">
                             {% if usuario.perfil_acesso == "GERENTE" %}
-                                <span class="inline-flex items-center rounded-full bg-blue-100 px-3 py-1 text-xs font-semibold text-blue-600">
+                                <span class="inline-flex items-center rounded-full bg-indigo-100 px-3 py-1 text-xs font-semibold text-indigo-600">
                                     Gerente
                                 </span>
                             {% elif usuario.perfil_acesso == "LIDER" %}
@@ -34,7 +34,7 @@
                                     Líder
                                 </span>
                             {% else %}
-                                <span class="inline-flex items-center rounded-full bg-gray-100 px-3 py-1 text-xs font-semibold text-gray-600">
+                                <span class="inline-flex items-center rounded-full bg-pink-100 px-3 py-1 text-xs font-semibold text-pink-600">
                                     Membro
                                 </span>
                             {% endif %}
@@ -56,31 +56,34 @@
                         <td class="px-4 py-3">
                             <div class="flex items-center justify-end gap-2">
                                 <a href="{% url 'usuarios:detalhe' usuario.pk %}"
-                                   class="inline-flex h-9 w-9 items-center justify-center rounded-full border border-gray-200 text-gray-500 transition hover:border-blue-200 hover:text-blue-600">
+                                   class="inline-flex h-9 w-9 items-center justify-center rounded-full border border-gray-200 text-gray-500 transition hover:border-indigo-200 hover:text-indigo-600">
                                     <i class="fas fa-eye"></i>
                                 </a>
                                 <a href="{% url 'usuarios:editar' usuario.pk %}"
-                                   class="inline-flex h-9 w-9 items-center justify-center rounded-full border border-blue-200 text-blue-500 transition hover:border-blue-300 hover:text-blue-600">
+                                   class="inline-flex h-9 w-9 items-center justify-center rounded-full border border-purple-200 text-purple-500 transition hover:border-purple-300 hover:text-purple-600">
                                     <i class="fas fa-pen"></i>
                                 </a>
-                                <form method="post" action="{% url 'usuarios:status' usuario.pk %}">
-                                    {% csrf_token %}
-                                    <input type="hidden" name="next" value="{{ request.get_full_path }}">
-                                    {% if usuario.ativo %}
-                                        <input type="hidden" name="acao" value="desativar">
-                                        <button type="submit"
-                                                class="inline-flex h-9 w-9 items-center justify-center rounded-full border border-red-200 text-red-500 transition hover:border-red-300 hover:text-red-600"
-                                                onclick="return confirm('Confirma desativar este usuário?')">
-                                            <i class="fas fa-user-slash"></i>
-                                        </button>
-                                    {% else %}
-                                        <input type="hidden" name="acao" value="ativar">
-                                        <button type="submit"
-                                                class="inline-flex h-9 w-9 items-center justify-center rounded-full border border-green-200 text-green-600 transition hover:border-green-300 hover:text-green-700">
-                                            <i class="fas fa-user-check"></i>
-                                        </button>
-                                    {% endif %}
-                                </form>
+                                {% if usuario.ativo %}
+                                    <button type="button"
+                                            class="inline-flex h-9 w-9 items-center justify-center rounded-full border border-red-200 text-red-500 transition hover:border-red-300 hover:text-red-600"
+                                            data-status-modal
+                                            data-url="{% url 'usuarios:status' usuario.pk %}"
+                                            data-acao="desativar"
+                                            data-nome="{{ usuario.nome_completo|escape }}"
+                                            data-next="{{ request.get_full_path|escape }}">
+                                        <i class="fas fa-user-slash"></i>
+                                    </button>
+                                {% else %}
+                                    <button type="button"
+                                            class="inline-flex h-9 w-9 items-center justify-center rounded-full border border-green-200 text-green-600 transition hover:border-green-300 hover:text-green-700"
+                                            data-status-modal
+                                            data-url="{% url 'usuarios:status' usuario.pk %}"
+                                            data-acao="ativar"
+                                            data-nome="{{ usuario.nome_completo|escape }}"
+                                            data-next="{{ request.get_full_path|escape }}">
+                                        <i class="fas fa-user-check"></i>
+                                    </button>
+                                {% endif %}
                             </div>
                         </td>
                     </tr>
