+++ apps/usuarios/templates/usuarios/componentes/modal_confirmacao_status.html
@@ -1,0 +1,39 @@
+<div id="statusModal" class="fixed inset-0 z-50 hidden items-center justify-center">
+    <div class="absolute inset-0 bg-slate-900/60 backdrop-blur-sm" data-status-modal-close></div>
+    <div class="relative mx-4 w-full max-w-md rounded-3xl border border-gray-200 bg-white p-6 shadow-2xl">
+        <form method="post" id="statusModalForm" class="space-y-6">
+            {% csrf_token %}
+            <input type="hidden" name="acao" id="statusModalAction">
+            <input type="hidden" name="next" id="statusModalNext">
+
+            <div class="flex items-start gap-3">
+                <span id="statusModalIcon"
+                      class="inline-flex h-12 w-12 items-center justify-center rounded-full bg-purple-100 text-purple-600">
+                    <i class="fas fa-user-shield text-lg"></i>
+                </span>
+                <div class="space-y-1">
+                    <p id="statusModalTitle" class="text-lg font-semibold text-slate-900">
+                        Atualizar status do usuário
+                    </p>
+                    <p id="statusModalMessage" class="text-sm text-gray-500">
+                        Confirme a ação desejada.
+                    </p>
+                </div>
+            </div>
+
+            <div class="flex justify-end gap-2">
+                <button type="button"
+                        class="inline-flex items-center gap-2 rounded-xl border border-gray-200 px-4 py-2 text-sm font-medium text-gray-600 transition hover:border-gray-300 hover:bg-gray-50"
+                        data-status-modal-close>
+                    Cancelar
+                </button>
+                <button type="submit"
+                        id="statusModalConfirm"
+                        class="inline-flex items-center gap-2 rounded-xl px-5 py-2 text-sm font-semibold text-white shadow transition focus:outline-none focus:ring-2 focus:ring-offset-1">
+                    <i class="fas fa-check-circle"></i>
+                    <span id="statusModalConfirmText">Confirmar</span>
+                </button>
+            </div>
+        </form>
+    </div>
+</div>
