+++ apps/usuarios/templates/usuarios/index.html
@@ -3,40 +3,52 @@
 {% block title %}Gestão de Usuários{% endblock %}
 
 {% block content %}
-<div class="max-w-6xl mx-auto space-y-6">
-    <nav aria-label="Breadcrumb" class="text-sm text-gray-400">
-        <ol class="flex items-center gap-2">
-            <li>
-                <a href="/" class="flex items-center gap-1 text-gray-400 hover:text-gray-600">
-                    <i class="fas fa-home"></i>
-                    <span class="hidden sm:inline">Athos Insight</span>
+<div class="mx-auto max-w-7xl px-4 py-6 space-y-6">
+    <section class="rounded-3xl border border-gray-200 bg-white shadow-sm">
+        <div class="px-6 py-6 space-y-6">
+            <nav class="flex items-center gap-2 text-sm text-gray-500">
+                <a href="/" class="inline-flex items-center gap-2 text-gray-500 transition hover:text-gray-700">
+                    <span class="inline-flex h-9 w-9 items-center justify-center rounded-xl border border-gray-200 bg-gray-50 text-gray-500 shadow-sm">
+                        <svg xmlns="http://www.w3.org/2000/svg" class="h-4 w-4" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+                            <rect x="3" y="3" width="7" height="7" rx="2" stroke-width="2" />
+                            <rect x="14" y="3" width="7" height="7" rx="2" stroke-width="2" />
+                            <rect x="14" y="14" width="7" height="7" rx="2" stroke-width="2" />
+                            <rect x="3" y="14" width="7" height="7" rx="2" stroke-width="2" />
+                        </svg>
+                    </span>
+                    <span class="font-medium tracking-wide">ATHOS INSIGHT</span>
                 </a>
-            </li>
-            <li class="text-gray-300">/</li>
-            <li>
-                <a href="{% url 'usuarios:lista' %}" class="text-gray-500 hover:text-gray-700">
-                    Gerenciamento de Usuários
-                </a>
-            </li>
-        </ol>
-    </nav>
+                <span class="text-gray-300">›</span>
+                <span class="text-gray-500">Usuários</span>
+                <span class="hidden text-gray-300 sm:inline">›</span>
+                <span class="hidden text-gray-700 sm:inline font-semibold">Gestão</span>
+            </nav>
 
-    <div class="flex flex-col md:flex-row md:items-center md:justify-between gap-4">
-        <div>
-            <h1 class="text-3xl font-semibold text-gray-800">Gerenciamento de Usuários</h1>
-            <p class="text-sm text-gray-500">Cadastre e gerencie usuários do sistema</p>
+            <div class="flex flex-col gap-4 md:flex-row md:items-center md:justify-between">
+                <div class="flex items-start gap-4">
+                    <button type="button"
+                            onclick="window.history.back()"
+                            class="inline-flex items-center gap-2 rounded-xl border border-gray-200 bg-white px-4 py-2 text-sm font-medium text-gray-600 transition hover:border-gray-300 hover:bg-gray-50">
+                        ← Voltar
+                    </button>
+                    <div>
+                        <h1 class="text-3xl font-semibold text-slate-900">Gerenciamento de Usuários</h1>
+                        <p class="mt-1 text-sm text-gray-500">Cadastre, filtre e acompanhe usuários do ecossistema Athos.</p>
+                    </div>
+                </div>
+                <a href="{% url 'usuarios:criar' %}"
+                   class="inline-flex items-center gap-2 rounded-xl bg-gradient-to-r from-indigo-500 via-purple-500 to-pink-500 px-5 py-2 text-sm font-semibold text-white shadow-md transition hover:brightness-105 focus:outline-none focus:ring-2 focus:ring-purple-200 focus:ring-offset-1">
+                    <i class="fas fa-plus text-white/90"></i>
+                    Novo Usuário
+                </a>
+            </div>
         </div>
-        <a href="{% url 'usuarios:criar' %}"
-           class="inline-flex items-center gap-2 rounded-full bg-blue-600 px-5 py-2 text-sm font-semibold text-white shadow-sm hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-blue-500 focus:ring-offset-2">
-            <i class="fas fa-plus"></i>
-            Novo Usuário
-        </a>
-    </div>
+    </section>
 
     {% if messages %}
         <div class="space-y-2">
             {% for message in messages %}
-                <div class="rounded-md border border-blue-200 bg-blue-50 px-4 py-3 text-sm text-blue-700">
+                <div class="rounded-xl border border-indigo-100 bg-indigo-50 px-4 py-3 text-sm font-medium text-indigo-700">
                     {{ message }}
                 </div>
             {% endfor %}
@@ -48,8 +60,8 @@
     <section class="space-y-4">
         <div class="flex items-center justify-between">
             <div>
-                <h2 class="text-lg font-semibold text-gray-700">Lista de Usuários</h2>
-                <p class="text-xs text-gray-400">
+                <h2 class="text-lg font-semibold text-slate-800">Lista de Usuários</h2>
+                <p class="text-xs uppercase tracking-wide text-gray-400">
                     {{ usuarios_count }} usuário(s) encontrado(s)
                 </p>
             </div>
@@ -58,4 +70,10 @@
         {% include "usuarios/componentes/lista.html" with usuarios=usuarios %}
     </section>
 </div>
+{% include "usuarios/componentes/modal_confirmacao_status.html" %}
+{% endblock %}
+
+{% block extra_js %}
+    {{ block.super }}
+    {% include "usuarios/componentes/modal_confirmacao_status_script.html" %}
 {% endblock %}
