+++ tests/unit/apps/usuarios/__init__.py
@@ -1,0 +1,1 @@
+"""Test package for usuarios app."""
