+++ tests/unit/apps/usuarios/test_models.py
@@ -1,0 +1,32 @@
+"""Unit tests for the custom Usuario model."""
+
+from django.contrib.auth import get_user_model
+from django.test import TestCase
+
+from apps.usuarios.models import ContratoChoices, PerfilAcessoChoices
+
+Usuario = get_user_model()
+
+
+class UsuarioModelTests(TestCase):
+    """Cenários focados no modelo customizado."""
+
+    def test_salvar_sincroniza_is_active(self):
+        usuario = Usuario.objects.create_user(
+            username="janedoe",
+            nome_completo="Jane Doe",
+            email="jane@example.com",
+            contrato=ContratoChoices.CLT,
+            cargo="Gerente de Projeto",
+            perfil_acesso=PerfilAcessoChoices.GERENTE,
+            password="Str0ng@123",
+        )
+        self.assertTrue(usuario.ativo)
+        self.assertTrue(usuario.is_active)
+
+        usuario.ativo = False
+        usuario.save()
+        usuario.refresh_from_db()
+
+        self.assertFalse(usuario.ativo)
+        self.assertFalse(usuario.is_active)
